requirements.txt
-e git+https://github.com/joshblum/django-with-audit.git#egg=django
Markdown==2.2.1
django-registration==0.8
django-registration-defaults==0.3
django-kronos
django-evolution